+++ p2panda-rs/src/test_utils/mocks/mod.rs
@@ -1,0 +1,8 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+
+//! Mock p2panda node and client for demo and testing purposes.
+
+pub mod logs;
+pub mod node;
+pub mod materialiser;
+pub mod client;